refact: split nostr api in order to make jest test work
